Calculate the variance component for number of times unit is a match.

This function calculates K_M_var_factor, a numeric vector. Each entry in this vector is a function of the number of times each unit is matched to, aka \(K_M(i)\) (corresponding to Kiw, where \(M\) corresponds to M_matches. The calculation in this function comes from Theorem 7, page 251 of Abadie and Imbens (2006) Econometrica. The K_M_var_factor is an important component in the variance estimation, and is called from estVarAI2006 in estimateTau.

calcKMVarFactor(Kiw, M_matches)

Arguments

Kiw
A vector of times each unit is matched to

M_matches
Number of matches per unit for imputing potential outcomes, as in Abadie and Imbens (2006).

Value

A numeric vector.

This function is exported for use in other packages.

References

Abadie, A., & Imbens, G. W. (2006). Large sample properties of matching estimators for average treatment effects. econometrica, 74(1), 235-267. https://doi.org/10.1111/j.1468-0262.2006.00655.x