Calculate Variance as in Abadie & Imbens (2016) JASA

This function calculates the estimated variance of matching estimator when matching on propensity scores that were estimated with multinomial logistic regression. This method takes into account the uncertainty from the treatment model. This function does no "heavy lifting" in that it only assembles pieces that have been estimated elsewhere (i.e., see estSigSq to see how the matching procedures are carried out).

calcSigSqAI2016(W, X_covars, Y, N, num_trts, trt_levels, match_mat_AI2016,
  prop_score_ests, vcov_coeff, tau_dfm, ...)

Arguments

W
A treatment vector (1 x n) with numerical values indicating treatment groups

X_covars
The matrix of covariates (not the matrix of estimated GPS)

Y
A response vector (1 x n)

N
The total number of units

num_trts
The number of unique treatments (3+), a scalar

trt_levels
A vector (of length num_trts providing the unique treatment levels

match_mat_AI2016
Information on matches from matchAllTreatments

prop_score_ests
The matrix of propensity scores

vcov_coeff
The variance-covariance (inverse fisher information) matrix from the multinomial logistic regression model

tau_dfm
The dataframe of estimates and estimated standard errors output from estimateTau

...
the dots argument

Value

A list with the updated tau_dfm includeing a column for VarianceAI2016, and a list object AI2016_args with potentially helpful extra information.

References

Abadie, A., & Imbens, G. W. (2016). Matching on the estimated propensity score. Econometrica, 84(2), 781-807. https://doi.org/10.3982/ECTA11293