Perform matching methods to estimate variance for one treatment level

This function executes the matching methods to estimate the components for estimating the variance (sigma squared) of the matching estimator. All matching methods (specified with match_on) will estimate the variance as in Abadie and Imbens (2006). When match_on = "multinom" this function will carry out additional matching procedures to estimate the variance as described in Abadie and Imbens (2016).

estSigSq(X, Y, which_same_trt, N_this_trt, var_options, match_on,
  var_args_AI2016)

Arguments

X
A covariate matrix (p x n) with no intercept. When match_on="existing", then X must be a vector (1 x n) of user-specified propensity scores.

Y
A response vector (1 x n)

which_same_trt
Numeric vector providing indices of individuals observed to have treatment level kk

N_this_trt
The number of units observed to have this treatment level.

var_options
Options for carrying out matching for variance estimation.

match_on
User specifies "covariates" to match on raw covariates, or "existing" to match on user-supplied propensity score values, or "polr" or "multinom" to fit a propensity score model.

var_args_AI2016
A list of arguments for carrying out matching procedures to estimate components in the VarianceAI2016 variance estimates (see calcSigSqAI2016).

Note that these variance components are combined in estimateTau (and perhaps calcSigSqAI2016).

Value

A list of one or two elements. sigsqiw_kk is a vector with the estimated variance component for units observed to have the kk^th treatment level. When match_on = "multinom", the list will also have an element for match_mat_AI2016_kk_two_cols, which is a Ntot-by-2 matrix with matching information that will eventually be passed to calcSigSqAI2016.

References

Abadie, A., & Imbens, G. W. (2006). Large sample properties of matching estimators for average treatment effects. econometrica, 74(1), 235-267. https://doi.org/10.1111/j.1468-0262.2006.00655.x

Abadie, A., & Imbens, G. W. (2016). Matching on the estimated propensity score. Econometrica, 84(2), 781-807. https://doi.org/10.3982/ECTA11293