Estimate Imputed Outcomes and AI06 Variance Components

Mostly calls a subfunction for each treatment level, which calls subfunctions that either carry out the matched imputation or the variance estimation.

matchAllTreatments(N, X, W, Y, num_trts, trt_levels, N_per_trt,
  unit_ids_sorted, M_matches, J_var_matches, match_on, ...)

Arguments

N
The total number of units

X
A covariate matrix (p x n) with no intercept. When match_on="existing", then X must be a vector (1 x n) of user-specified propensity scores.

W
A treatment vector (1 x n) with numerical values indicating treatment groups

Y
A response vector (1 x n)

num_trts
The number of unique treatments (3+), a scalar

trt_levels
A vector (of length num_trts providing the unique treatment levels

N_per_trt
A vector (of length num_trts) indicating the number of units observed to have each treatment level

unit_ids_sorted
The unit identifiers in correct, sorted order.

M_matches
Number of matches per unit for imputing potential outcomes, as in Abadie and Imbens (2006).

J_var_matches
Number of matches when estimating \(\sigma^2(X,W)\) as in Abadie and Imbens (2006).

match_on
User specifies "covariates" to match on raw covariates, or "existing" to match on user-supplied propensity score values, or "polr" or "multinom" to fit a propensity score model.

...
the dots argument

Value

A list including at most:

Yiw: A matrix of all imputed (or observed) potential outcomes for each unit

mean_Yiw: A vector of the average (across all units) of estimated/imputed potential outcomes

sigsqiw: Estimated variance components (from Abadie and Imbens (2006)'s method) for each unit

Kiw: The vector of number of times unit i used as a match

impute_match_data: extra information from the main matching procedure

match_mat_AI2016: When match_on=`multinom` this additional information will be output for calcSigSqAI2016 (from Abadie and Imbens (2016)'s method)

A few of the necessary arguments are output from the reorderByTreatment function.

References

Yang, S., Imbens G. W., Cui, Z., Faries, D. E., & Kadziola, Z. (2016) Propensity Score Matching and Subclassification in Observational Studies with Multi-Level Treatments. Biometrics, 72, 1055-1065. https://doi.org/10.1111/biom.12505

Abadie, A., & Imbens, G. W. (2006). Large sample properties of matching estimators for average treatment effects. econometrica, 74(1), 235-267. https://doi.org/10.1111/j.1468-0262.2006.00655.x

Abadie, A., & Imbens, G. W. (2016). Matching on the estimated propensity score. Econometrica, 84(2), 781-807. https://doi.org/10.3982/ECTA11293