Matching Estimators for Multiple Treatments from Yang et al (2016).

This function carries out matching on covariates or on propensity scores, and is similar to the 'legacy' functions multilevelMatchX and multilevelGPSMatch.

multiMatch(Y, W, X, trimming = NULL, match_on,
  model_options = list(reference_level = sort(W)[1]), M_matches = 1,
  J_var_matches = 1)

Arguments

Y
A response vector (1 x n)

W
A treatment vector (1 x n) with numerical values indicating treatment groups

X
A covariate matrix (p x n) with no intercept. When match_on="existing", then X must be a vector (1 x n) of user-specified propensity scores.

trimming
an indicator of whether trimming the sample to ensure overlap

match_on
"multinom", "polr", "existing", or "covariates",

model_options
A list of the options to pass to propensity model. Currently under development. Can only pass reference level to multinomial logisitc regression.

M_matches
Number of matches per unit for imputing potential outcomes, as in Abadie and Imbens (2006). Currently can only support M=1.

J_var_matches
Number of matches when estimating sigmasq(X,W) as in Abadie and Imbens (2006). Currently can only support J=1.

Value

according to estimateTau, including at most:

tauestimate: a vector of estimates for pairwise treatment effects

varestimate: a vector of variance estimates for tauestimate, using Abadie and Imbens (2006)'s method

varestimateAI2016: a vector of variance estimates for tauestimate, when matching on the generalized propensity score, using Abadie & Imbens (2016)'s method. This variance estimate takes into account of the uncertainty in estimating the GPS.

analysis_idx: a list containing the indices_kept (analyzed) and indices_dropped (trimmed) based on Crump et al. (2009)'s method.

References

Yang, S., Imbens G. W., Cui, Z., Faries, D. E., & Kadziola, Z. (2016) Propensity Score Matching and Subclassification in Observational Studies with Multi-Level Treatments. Biometrics, 72, 1055-1065. https://doi.org/10.1111/biom.12505

Abadie, A., & Imbens, G. W. (2006). Large sample properties of matching estimators for average treatment effects. econometrica, 74(1), 235-267. https://doi.org/10.1111/j.1468-0262.2006.00655.x

Abadie, A., & Imbens, G. W. (2016). Matching on the estimated propensity score. Econometrica, 84(2), 781-807. https://doi.org/10.3982/ECTA11293

Crump, R. K., Hotz, V. J., Imbens, G. W., & Mitnik, O. A. (2009). Dealing with limited overlap in estimation of average treatment effects. Biometrika, 96(1), 187-199. https://doi.org/10.1093/biomet/asn055

See also

Examples

  X <- as.matrix(c(5.5,10.6,3.1,8.7,5.1,10.2,9.8,4.4,4.9),ncol=1)
  Y <- c(102,105,120,130,100,80,94,108,96)
  W <- c(1,1,1,3,2,3,2,1,2)
  multiMatch(Y,W,X,trimming=0,match_on="multinom")

#> no unit IDs supplied; unit_ids will be assigned generically

#> $results
#>         Param Trt1 Trt2   Estimate   Variance VarianceAI2016
#> 1 EY(2)-EY(1)    1    2 -10.444444   8.545953       8.302024
#> 2 EY(3)-EY(1)    1    3   6.666667 616.913580     411.456234
#> 3 EY(3)-EY(2)    2    3  17.111111 611.122085     434.247037
#> 
#> $analysis_idx
#> NULL
#> 
#> $mu
#>   Param Trt  Estimate
#> 1 EY(1)   1 106.66667
#> 2 EY(2)   2  96.22222
#> 3 EY(3)   3 113.33333
#> 
#> $impute_mat
#>          1   2   3
#> unit_1 102 100 130
#> unit_2 105  94  80
#> unit_3 120  96 130
#> unit_4 105  96 130
#> unit_5 102 100 130
#> unit_6 105  94  80
#> unit_7 105  94  80
#> unit_8 108  96 130
#> unit_9 108  96 130
#> 
#> $estimate_args
#> $estimate_args$W
#> unit_1 unit_2 unit_3 unit_8 unit_5 unit_7 unit_9 unit_4 unit_6 
#>      1      1      1      1      2      2      2      3      3 
#> 
#> $estimate_args$X
#>                1         2          3
#> unit_1 0.5544505 0.3829898 0.06255971
#> unit_2 0.2007354 0.2444163 0.55484828
#> unit_3 0.6434724 0.3404115 0.01611607
#> unit_8 0.5995704 0.3664938 0.03393583
#> unit_5 0.5719138 0.3778737 0.05021253
#> unit_7 0.2640193 0.2941204 0.44186033
#> unit_9 0.5801676 0.3749000 0.04493242
#> unit_4 0.3539403 0.3489165 0.29714313
#> unit_6 0.2317252 0.2698803 0.49839443
#> 
#> $estimate_args$Y
#> unit_1 unit_2 unit_3 unit_8 unit_5 unit_7 unit_9 unit_4 unit_6 
#>    102    105    120    108    100     94     96    130     80 
#> 
#> $estimate_args$N
#> [1] 9
#> 
#> $estimate_args$num_trts
#> [1] 3
#> 
#> $estimate_args$trt_levels
#> [1] 1 2 3
#> 
#> $estimate_args$N_per_trt
#> W
#> 1 2 3 
#> 4 3 2 
#> 
#> $estimate_args$num_contrasts
#> [1] 3
#> 
#> $estimate_args$orig_to_sorted
#> [1] 1 2 3 8 5 7 9 4 6
#> 
#> $estimate_args$sorted_to_orig
#> [1] 1 2 3 8 5 9 6 4 7
#> 
#> $estimate_args$unit_ids_unsorted
#> [1] "unit_1" "unit_2" "unit_3" "unit_4" "unit_5" "unit_6" "unit_7" "unit_8"
#> [9] "unit_9"
#> 
#> $estimate_args$unit_ids_sorted
#> [1] "unit_1" "unit_2" "unit_3" "unit_8" "unit_5" "unit_7" "unit_9" "unit_4"
#> [9] "unit_6"
#> 
#> $estimate_args$match_on
#> [1] "multinom"
#> 
#> $estimate_args$model_options
#> $estimate_args$model_options$reference_level
#> [1] 1
#> 
#> 
#> $estimate_args$M_matches
#> [1] 1
#> 
#> $estimate_args$J_var_matches
#> [1] 1
#> 
#> $estimate_args$Yiw
#>          1   2   3
#> unit_1 102 100 130
#> unit_2 105  94  80
#> unit_3 120  96 130
#> unit_8 108  96 130
#> unit_5 102 100 130
#> unit_7 105  94  80
#> unit_9 108  96 130
#> unit_4 105  96 130
#> unit_6 105  94  80
#> 
#> $estimate_args$mean_Yiw
#>         1         2         3 
#> 106.66667  96.22222 113.33333 
#> 
#> $estimate_args$Kiw
#>        Kiw
#> unit_1   1
#> unit_2   3
#> unit_3   0
#> unit_8   1
#> unit_5   1
#> unit_7   2
#> unit_9   3
#> unit_4   5
#> unit_6   2
#> 
#> $estimate_args$impute_match_data
#> $estimate_args$impute_match_data[[1]]
#> $estimate_args$impute_match_data[[1]]$Y
#>  [1] 100  94  96 130  80 102 105 108 105 105
#> 
#> $estimate_args$impute_match_data[[1]]$Tr
#>  [1] 1 1 1 1 1 0 0 0 0 0
#> 
#> $estimate_args$impute_match_data[[1]]$X
#>         unit_5    unit_7    unit_9    unit_4    unit_6
#> [1,] 0.5719138 0.2640193 0.5801676 0.3539403 0.2317252
#> [2,] 0.5544505 0.2007354 0.5995704 0.2007354 0.2007354
#> 
#> $estimate_args$impute_match_data[[1]]$orig.weighted.treated.nobs
#> [1] 5
#> 
#> 
#> $estimate_args$impute_match_data[[2]]
#> $estimate_args$impute_match_data[[2]]$Y
#>  [1] 102 105 120 108 130  80 100  94  96  96  96  94
#> 
#> $estimate_args$impute_match_data[[2]]$Tr
#>  [1] 1 1 1 1 1 1 0 0 0 0 0 0
#> 
#> $estimate_args$impute_match_data[[2]]$X
#>         unit_1    unit_2    unit_3    unit_8    unit_4    unit_6
#> [1,] 0.3829898 0.2444163 0.3404115 0.3664938 0.3489165 0.2698803
#> [2,] 0.3778737 0.2941204 0.3749000 0.3749000 0.3749000 0.2941204
#> 
#> $estimate_args$impute_match_data[[2]]$orig.weighted.treated.nobs
#> [1] 6
#> 
#> 
#> $estimate_args$impute_match_data[[3]]
#> $estimate_args$impute_match_data[[3]]$Y
#>  [1] 102 105 120 108 100  94  96 130  80 130 130 130  80 130
#> 
#> $estimate_args$impute_match_data[[3]]$Tr
#>  [1] 1 1 1 1 1 1 1 0 0 0 0 0 0 0
#> 
#> $estimate_args$impute_match_data[[3]]$X
#>          unit_1    unit_2     unit_3     unit_8     unit_5    unit_7     unit_9
#> [1,] 0.06255971 0.5548483 0.01611607 0.03393583 0.05021253 0.4418603 0.04493242
#> [2,] 0.29714313 0.4983944 0.29714313 0.29714313 0.29714313 0.4983944 0.29714313
#> 
#> $estimate_args$impute_match_data[[3]]$orig.weighted.treated.nobs
#> [1] 7
#> 
#> 
#> 
#> $estimate_args$sigsqiw
#>        sigsqiw
#> unit_1    18.0
#> unit_2     4.5
#> unit_3    72.0
#> unit_8    72.0
#> unit_5     8.0
#> unit_7     2.0
#> unit_9     8.0
#> unit_4  1250.0
#> unit_6  1250.0
#> 
#> $estimate_args$match_mat_AI2016
#>       m1.1 m1.2 m2.1 m2.2 m3.1 m3.2
#>  [1,]    1    4    5    7    8    9
#>  [2,]    2    1    6    7    8    9
#>  [3,]    3    4    5    7    8    9
#>  [4,]    4    3    5    7    8    9
#>  [5,]    1    4    5    7    8    9
#>  [6,]    1    2    6    7    8    9
#>  [7,]    1    4    7    5    8    9
#>  [8,]    1    2    5    7    8    9
#>  [9,]    1    2    6    7    9    8
#> 
#> 
#> $model
#> Call:
#> nnet::multinom(formula = W.ref ~ X)
#> 
#> Coefficients:
#>   (Intercept)         X
#> 2  -0.9812827 0.1111479
#> 3  -5.6312867 0.6271692
#> 
#> Residual Deviance: 16.47185 
#> AIC: 24.47185 
#> 
#> $propensity_scores
#>                1         2          3
#> unit_1 0.5544505 0.3829898 0.06255971
#> unit_2 0.2007354 0.2444163 0.55484828
#> unit_3 0.6434724 0.3404115 0.01611607
#> unit_8 0.5995704 0.3664938 0.03393583
#> unit_5 0.5719138 0.3778737 0.05021253
#> unit_7 0.2640193 0.2941204 0.44186033
#> unit_9 0.5801676 0.3749000 0.04493242
#> unit_4 0.3539403 0.3489165 0.29714313
#> unit_6 0.2317252 0.2698803 0.49839443
#> 
#> $impute_match_data
#> $impute_match_data[[1]]
#> $impute_match_data[[1]]$Y
#>  [1] 100  94  96 130  80 102 105 108 105 105
#> 
#> $impute_match_data[[1]]$Tr
#>  [1] 1 1 1 1 1 0 0 0 0 0
#> 
#> $impute_match_data[[1]]$X
#>         unit_5    unit_7    unit_9    unit_4    unit_6
#> [1,] 0.5719138 0.2640193 0.5801676 0.3539403 0.2317252
#> [2,] 0.5544505 0.2007354 0.5995704 0.2007354 0.2007354
#> 
#> $impute_match_data[[1]]$orig.weighted.treated.nobs
#> [1] 5
#> 
#> 
#> $impute_match_data[[2]]
#> $impute_match_data[[2]]$Y
#>  [1] 102 105 120 108 130  80 100  94  96  96  96  94
#> 
#> $impute_match_data[[2]]$Tr
#>  [1] 1 1 1 1 1 1 0 0 0 0 0 0
#> 
#> $impute_match_data[[2]]$X
#>         unit_1    unit_2    unit_3    unit_8    unit_4    unit_6
#> [1,] 0.3829898 0.2444163 0.3404115 0.3664938 0.3489165 0.2698803
#> [2,] 0.3778737 0.2941204 0.3749000 0.3749000 0.3749000 0.2941204
#> 
#> $impute_match_data[[2]]$orig.weighted.treated.nobs
#> [1] 6
#> 
#> 
#> $impute_match_data[[3]]
#> $impute_match_data[[3]]$Y
#>  [1] 102 105 120 108 100  94  96 130  80 130 130 130  80 130
#> 
#> $impute_match_data[[3]]$Tr
#>  [1] 1 1 1 1 1 1 1 0 0 0 0 0 0 0
#> 
#> $impute_match_data[[3]]$X
#>          unit_1    unit_2     unit_3     unit_8     unit_5    unit_7     unit_9
#> [1,] 0.06255971 0.5548483 0.01611607 0.03393583 0.05021253 0.4418603 0.04493242
#> [2,] 0.29714313 0.4983944 0.29714313 0.29714313 0.29714313 0.4983944 0.29714313
#> 
#> $impute_match_data[[3]]$orig.weighted.treated.nobs
#> [1] 7
#> 
#> 
#> 
#> $impute_mat_sorted
#>          1   2   3
#> unit_1 102 100 130
#> unit_2 105  94  80
#> unit_3 120  96 130
#> unit_8 108  96 130
#> unit_5 102 100 130
#> unit_7 105  94  80
#> unit_9 108  96 130
#> unit_4 105  96 130
#> unit_6 105  94  80
#> 
#> $AI2016_args
#> $AI2016_args$c_vector
#>      [,1]        [,2] [,3]        [,4]
#> [1,]    0  0.04536431    0  -1.4222059
#> [2,]    0 -1.02393228    0   0.7915606
#> [3,]    0 12.50001000    0 -29.1666552
#> 
#> $AI2016_args$vcov_coeff
#>               2:(Intercept)         2:X 3:(Intercept)         3:X
#> 2:(Intercept)     4.2775709 -0.58729558     1.9261788 -0.27779100
#> 2:X              -0.5872956  0.09365331    -0.2826604  0.04689727
#> 3:(Intercept)     1.9261788 -0.28266040    19.5121429 -2.10608615
#> 3:X              -0.2777910  0.04689727    -2.1060861  0.24229598
#> 
#> $AI2016_args$c_matrix
#>       [,1]      [,2] [,3]      [,4]
#>  [1,]    0 14.362119    0 -35.15401
#>  [2,]    0  9.165612    0 -16.69319
#>  [3,]    0 12.765431    0 -36.89565
#>  [4,]    0 13.743518    0 -36.22741
#>  [5,]    0 14.170264    0 -35.61703
#>  [6,]    0 11.029515    0 -20.93024
#>  [7,]    0 14.058749    0 -35.81503
#>  [8,]    0 13.084370    0 -26.35713
#>  [9,]    0 10.120512    0 -18.81021
#> 
#> 

  multiMatch(Y,W,X,trimming=1,match_on="multinom")

#> no unit IDs supplied; unit_ids will be assigned generically

#> $results
#>         Param Trt1 Trt2 Estimate   Variance VarianceAI2016
#> 1 EY(2)-EY(1)    1    2   -9.375   7.794922       5.072057
#> 2 EY(3)-EY(1)    1    3    5.875 582.654297     383.848575
#> 3 EY(3)-EY(2)    2    3   15.250 576.304688     430.978089
#> 
#> $analysis_idx
#> $analysis_idx$indices_kept
#> unit_1 unit_2 unit_4 unit_5 unit_6 unit_7 unit_8 unit_9 
#>      1      2      4      5      6      7      8      9 
#> 
#> $analysis_idx$indices_dropped
#> [1] 3
#> 
#> $analysis_idx$ids_kept
#> [1] "unit_1" "unit_2" "unit_4" "unit_5" "unit_6" "unit_7" "unit_8" "unit_9"
#> 
#> $analysis_idx$ids_dropped
#> [1] "unit_3"
#> 
#> 
#> $mu
#>   Param Trt Estimate
#> 1 EY(1)   1  105.375
#> 2 EY(2)   2   96.000
#> 3 EY(3)   3  111.250
#> 
#> $impute_mat
#>          1   2   3
#> unit_1 102 100 130
#> unit_2 105  94  80
#> unit_4 102  94 130
#> unit_5 108 100 130
#> unit_6 105  94  80
#> unit_7 105  94  80
#> unit_8 108  96 130
#> unit_9 108  96 130
#> 
#> $estimate_args
#> $estimate_args$W
#> unit_1 unit_2 unit_8 unit_5 unit_7 unit_9 unit_4 unit_6 
#>      1      1      1      2      2      2      3      3 
#> 
#> $estimate_args$X
#>                1         2          3
#> unit_1 0.4537905 0.4777830 0.06842650
#> unit_2 0.2432034 0.2063156 0.55048103
#> unit_8 0.4572150 0.5043475 0.03843750
#> unit_5 0.4560403 0.4883558 0.05560389
#> unit_7 0.2978420 0.2613750 0.44078293
#> unit_9 0.4567022 0.4932251 0.05007276
#> unit_4 0.3643634 0.3350017 0.30063489
#> unit_6 0.2707789 0.2336334 0.49558769
#> 
#> $estimate_args$Y
#> unit_1 unit_2 unit_8 unit_5 unit_7 unit_9 unit_4 unit_6 
#>    102    105    108    100     94     96    130     80 
#> 
#> $estimate_args$N
#> [1] 8
#> 
#> $estimate_args$num_trts
#> [1] 3
#> 
#> $estimate_args$trt_levels
#> [1] 1 2 3
#> 
#> $estimate_args$N_per_trt
#> W
#> 1 2 3 
#> 3 3 2 
#> 
#> $estimate_args$num_contrasts
#> [1] 3
#> 
#> $estimate_args$orig_to_sorted
#> [1] 1 2 7 4 6 8 3 5
#> 
#> $estimate_args$sorted_to_orig
#> [1] 1 2 7 4 8 5 3 6
#> 
#> $estimate_args$unit_ids_unsorted
#> [1] "unit_1" "unit_2" "unit_4" "unit_5" "unit_6" "unit_7" "unit_8" "unit_9"
#> 
#> $estimate_args$unit_ids_sorted
#> [1] "unit_1" "unit_2" "unit_8" "unit_5" "unit_7" "unit_9" "unit_4" "unit_6"
#> 
#> $estimate_args$analysis_idx
#> $estimate_args$analysis_idx$indices_kept
#> unit_1 unit_2 unit_4 unit_5 unit_6 unit_7 unit_8 unit_9 
#>      1      2      4      5      6      7      8      9 
#> 
#> $estimate_args$analysis_idx$indices_dropped
#> [1] 3
#> 
#> $estimate_args$analysis_idx$ids_kept
#> [1] "unit_1" "unit_2" "unit_4" "unit_5" "unit_6" "unit_7" "unit_8" "unit_9"
#> 
#> $estimate_args$analysis_idx$ids_dropped
#> [1] "unit_3"
#> 
#> 
#> $estimate_args$match_on
#> [1] "multinom"
#> 
#> $estimate_args$model_options
#> $estimate_args$model_options$reference_level
#> [1] 1
#> 
#> 
#> $estimate_args$M_matches
#> [1] 1
#> 
#> $estimate_args$J_var_matches
#> [1] 1
#> 
#> $estimate_args$Yiw
#>          1   2   3
#> unit_1 102 100 130
#> unit_2 105  94  80
#> unit_8 108  96 130
#> unit_5 108 100 130
#> unit_7 105  94  80
#> unit_9 108  96 130
#> unit_4 102  94 130
#> unit_6 105  94  80
#> 
#> $estimate_args$mean_Yiw
#>       1       2       3 
#> 105.375  96.000 111.250 
#> 
#> $estimate_args$Kiw
#>        Kiw
#> unit_1   1
#> unit_2   2
#> unit_8   2
#> unit_5   1
#> unit_7   3
#> unit_9   1
#> unit_4   4
#> unit_6   2
#> 
#> $estimate_args$impute_match_data
#> $estimate_args$impute_match_data[[1]]
#> $estimate_args$impute_match_data[[1]]$Y
#>  [1] 100  94  96 130  80 108 105 108 102 105
#> 
#> $estimate_args$impute_match_data[[1]]$Tr
#>  [1] 1 1 1 1 1 0 0 0 0 0
#> 
#> $estimate_args$impute_match_data[[1]]$X
#>         unit_5    unit_7    unit_9    unit_4    unit_6
#> [1,] 0.4560403 0.2978420 0.4567022 0.3643634 0.2707789
#> [2,] 0.4572150 0.2432034 0.4572150 0.4537905 0.2432034
#> 
#> $estimate_args$impute_match_data[[1]]$orig.weighted.treated.nobs
#> [1] 5
#> 
#> 
#> $estimate_args$impute_match_data[[2]]
#> $estimate_args$impute_match_data[[2]]$Y
#>  [1] 102 105 108 130  80 100  94  96  94  94
#> 
#> $estimate_args$impute_match_data[[2]]$Tr
#>  [1] 1 1 1 1 1 0 0 0 0 0
#> 
#> $estimate_args$impute_match_data[[2]]$X
#>         unit_1    unit_2    unit_8    unit_4    unit_6
#> [1,] 0.4777830 0.2063156 0.5043475 0.3350017 0.2336334
#> [2,] 0.4883558 0.2613750 0.4932251 0.2613750 0.2613750
#> 
#> $estimate_args$impute_match_data[[2]]$orig.weighted.treated.nobs
#> [1] 5
#> 
#> 
#> $estimate_args$impute_match_data[[3]]
#> $estimate_args$impute_match_data[[3]]$Y
#>  [1] 102 105 108 100  94  96 130  80 130 130  80 130
#> 
#> $estimate_args$impute_match_data[[3]]$Tr
#>  [1] 1 1 1 1 1 1 0 0 0 0 0 0
#> 
#> $estimate_args$impute_match_data[[3]]$X
#>         unit_1    unit_2    unit_8     unit_5    unit_7     unit_9
#> [1,] 0.0684265 0.5504810 0.0384375 0.05560389 0.4407829 0.05007276
#> [2,] 0.3006349 0.4955877 0.3006349 0.30063489 0.4955877 0.30063489
#> 
#> $estimate_args$impute_match_data[[3]]$orig.weighted.treated.nobs
#> [1] 6
#> 
#> 
#> 
#> $estimate_args$sigsqiw
#>        sigsqiw
#> unit_1    18.0
#> unit_2     4.5
#> unit_8    18.0
#> unit_5     8.0
#> unit_7    18.0
#> unit_9     8.0
#> unit_4  1250.0
#> unit_6  1250.0
#> 
#> $estimate_args$match_mat_AI2016
#>      m1.1 m1.2 m2.1 m2.2 m3.1 m3.2
#> [1,]    1    3    4    6    7    8
#> [2,]    2    1    4    5    7    8
#> [3,]    3    1    4    6    7    8
#> [4,]    1    3    4    6    7    8
#> [5,]    1    2    5    4    7    8
#> [6,]    1    3    6    4    7    8
#> [7,]    1    3    4    5    7    8
#> [8,]    1    2    4    5    8    7
#> 
#> 
#> $model
#> Call:
#> nnet::multinom(formula = W.ref ~ X)
#> 
#> Coefficients:
#>   (Intercept)           X
#> 2   0.2844751 -0.04235532
#> 3  -4.8130981  0.53113136
#> 
#> Residual Deviance: 15.3115 
#> AIC: 23.3115 
#> 
#> $propensity_scores
#>                1         2          3
#> unit_1 0.4537905 0.4777830 0.06842650
#> unit_2 0.2432034 0.2063156 0.55048103
#> unit_8 0.4572150 0.5043475 0.03843750
#> unit_5 0.4560403 0.4883558 0.05560389
#> unit_7 0.2978420 0.2613750 0.44078293
#> unit_9 0.4567022 0.4932251 0.05007276
#> unit_4 0.3643634 0.3350017 0.30063489
#> unit_6 0.2707789 0.2336334 0.49558769
#> 
#> $impute_match_data
#> $impute_match_data[[1]]
#> $impute_match_data[[1]]$Y
#>  [1] 100  94  96 130  80 108 105 108 102 105
#> 
#> $impute_match_data[[1]]$Tr
#>  [1] 1 1 1 1 1 0 0 0 0 0
#> 
#> $impute_match_data[[1]]$X
#>         unit_5    unit_7    unit_9    unit_4    unit_6
#> [1,] 0.4560403 0.2978420 0.4567022 0.3643634 0.2707789
#> [2,] 0.4572150 0.2432034 0.4572150 0.4537905 0.2432034
#> 
#> $impute_match_data[[1]]$orig.weighted.treated.nobs
#> [1] 5
#> 
#> 
#> $impute_match_data[[2]]
#> $impute_match_data[[2]]$Y
#>  [1] 102 105 108 130  80 100  94  96  94  94
#> 
#> $impute_match_data[[2]]$Tr
#>  [1] 1 1 1 1 1 0 0 0 0 0
#> 
#> $impute_match_data[[2]]$X
#>         unit_1    unit_2    unit_8    unit_4    unit_6
#> [1,] 0.4777830 0.2063156 0.5043475 0.3350017 0.2336334
#> [2,] 0.4883558 0.2613750 0.4932251 0.2613750 0.2613750
#> 
#> $impute_match_data[[2]]$orig.weighted.treated.nobs
#> [1] 5
#> 
#> 
#> $impute_match_data[[3]]
#> $impute_match_data[[3]]$Y
#>  [1] 102 105 108 100  94  96 130  80 130 130  80 130
#> 
#> $impute_match_data[[3]]$Tr
#>  [1] 1 1 1 1 1 1 0 0 0 0 0 0
#> 
#> $impute_match_data[[3]]$X
#>         unit_1    unit_2    unit_8     unit_5    unit_7     unit_9
#> [1,] 0.0684265 0.5504810 0.0384375 0.05560389 0.4407829 0.05007276
#> [2,] 0.3006349 0.4955877 0.3006349 0.30063489 0.4955877 0.30063489
#> 
#> $impute_match_data[[3]]$orig.weighted.treated.nobs
#> [1] 6
#> 
#> 
#> 
#> $impute_mat_sorted
#>          1   2   3
#> unit_1 102 100 130
#> unit_2 105  94  80
#> unit_8 108  96 130
#> unit_5 108 100 130
#> unit_7 105  94  80
#> unit_9 108  96 130
#> unit_4 102  94 130
#> unit_6 105  94  80
#> 
#> $AI2016_args
#> $AI2016_args$c_vector
#>      [,1]      [,2] [,3]       [,4]
#> [1,]    0  0.277578    0  -1.210117
#> [2,]    0 -5.121661    0   3.139818
#> [3,]    0 14.062674    0 -28.124873
#> 
#> $AI2016_args$vcov_coeff
#>               2:(Intercept)         2:X 3:(Intercept)         3:X
#> 2:(Intercept)     6.1334608 -0.81407641     3.0403215 -0.39723734
#> 2:X              -0.8140764  0.12126318    -0.3960346  0.05811916
#> 3:(Intercept)     3.0403215 -0.39603457    21.2168440 -2.27970447
#> 3:X              -0.3972373  0.05811916    -2.2797045  0.25886828
#> 
#> $AI2016_args$c_matrix
#>      [,1]      [,2] [,3]      [,4]
#> [1,]    0 17.916862    0 -34.93401
#> [2,]    0  7.736834    0 -16.85696
#> [3,]    0 18.913031    0 -36.05859
#> [4,]    0 18.313342    0 -35.41485
#> [5,]    0  9.801564    0 -20.97064
#> [6,]    0 18.495940    0 -35.62227
#> [7,]    0 12.562564    0 -26.22619
#> [8,]    0  8.761254    0 -18.91546
#> 
#> 

  multiMatch(Y,W,X,trimming=0,match_on="polr")

#> no unit IDs supplied; unit_ids will be assigned generically

#> $results
#>         Param Trt1 Trt2   Estimate   Variance VarianceAI2016
#> 1 EY(2)-EY(1)    1    2 -10.666667   9.111111             NA
#> 2 EY(3)-EY(1)    1    3   6.666667 615.580247             NA
#> 3 EY(3)-EY(2)    2    3  17.333333 613.925926             NA
#> 
#> $analysis_idx
#> NULL
#> 
#> $mu
#>   Param Trt Estimate
#> 1 EY(1)   1 106.6667
#> 2 EY(2)   2  96.0000
#> 3 EY(3)   3 113.3333
#> 
#> $impute_mat
#>          1   2   3
#> unit_1 102 100 130
#> unit_2 105  94  80
#> unit_3 120  96 130
#> unit_4 105  94 130
#> unit_5 102 100 130
#> unit_6 105  94  80
#> unit_7 105  94  80
#> unit_8 108  96 130
#> unit_9 108  96 130
#> 
#> $estimate_args
#> $estimate_args$W
#> unit_1 unit_2 unit_3 unit_8 unit_5 unit_7 unit_9 unit_4 unit_6 
#>      1      1      1      1      2      2      2      3      3 
#> 
#> $estimate_args$X
#>                1         2          3
#> unit_1 0.5345909 0.3390333 0.12637579
#> unit_2 0.1483133 0.3634129 0.48827384
#> unit_3 0.7362055 0.2076023 0.05619221
#> unit_8 0.6330845 0.2790739 0.08784167
#> unit_5 0.5711498 0.3179273 0.11092282
#> unit_7 0.1896979 0.3951802 0.41512190
#> unit_9 0.5891672 0.3069985 0.10383431
#> unit_4 0.2601697 0.4189477 0.32088265
#> unit_6 0.1679910 0.3805735 0.45143553
#> 
#> $estimate_args$Y
#> unit_1 unit_2 unit_3 unit_8 unit_5 unit_7 unit_9 unit_4 unit_6 
#>    102    105    120    108    100     94     96    130     80 
#> 
#> $estimate_args$N
#> [1] 9
#> 
#> $estimate_args$num_trts
#> [1] 3
#> 
#> $estimate_args$trt_levels
#> [1] 1 2 3
#> 
#> $estimate_args$N_per_trt
#> W
#> 1 2 3 
#> 4 3 2 
#> 
#> $estimate_args$num_contrasts
#> [1] 3
#> 
#> $estimate_args$orig_to_sorted
#> [1] 1 2 3 8 5 7 9 4 6
#> 
#> $estimate_args$sorted_to_orig
#> [1] 1 2 3 8 5 9 6 4 7
#> 
#> $estimate_args$unit_ids_unsorted
#> [1] "unit_1" "unit_2" "unit_3" "unit_4" "unit_5" "unit_6" "unit_7" "unit_8"
#> [9] "unit_9"
#> 
#> $estimate_args$unit_ids_sorted
#> [1] "unit_1" "unit_2" "unit_3" "unit_8" "unit_5" "unit_7" "unit_9" "unit_4"
#> [9] "unit_6"
#> 
#> $estimate_args$match_on
#> [1] "polr"
#> 
#> $estimate_args$model_options
#> $estimate_args$model_options$reference_level
#> [1] 1
#> 
#> 
#> $estimate_args$M_matches
#> [1] 1
#> 
#> $estimate_args$J_var_matches
#> [1] 1
#> 
#> $estimate_args$Yiw
#>          1   2   3
#> unit_1 102 100 130
#> unit_2 105  94  80
#> unit_3 120  96 130
#> unit_8 108  96 130
#> unit_5 102 100 130
#> unit_7 105  94  80
#> unit_9 108  96 130
#> unit_4 105  94 130
#> unit_6 105  94  80
#> 
#> $estimate_args$mean_Yiw
#>        1        2        3 
#> 106.6667  96.0000 113.3333 
#> 
#> $estimate_args$Kiw
#>        Kiw
#> unit_1   1
#> unit_2   3
#> unit_3   0
#> unit_8   1
#> unit_5   1
#> unit_7   3
#> unit_9   2
#> unit_4   5
#> unit_6   2
#> 
#> $estimate_args$impute_match_data
#> $estimate_args$impute_match_data[[1]]
#> $estimate_args$impute_match_data[[1]]$Y
#>  [1] 100  94  96 130  80 102 105 108 105 105
#> 
#> $estimate_args$impute_match_data[[1]]$Tr
#>  [1] 1 1 1 1 1 0 0 0 0 0
#> 
#> $estimate_args$impute_match_data[[1]]$X
#>         unit_5    unit_7    unit_9    unit_4    unit_6
#> [1,] 0.5711498 0.1896979 0.5891672 0.2601697 0.1679910
#> [2,] 0.5345909 0.1483133 0.6330845 0.1483133 0.1483133
#> 
#> $estimate_args$impute_match_data[[1]]$orig.weighted.treated.nobs
#> [1] 5
#> 
#> 
#> $estimate_args$impute_match_data[[2]]
#> $estimate_args$impute_match_data[[2]]$Y
#>  [1] 102 105 120 108 130  80 100  94  96  96  94  94
#> 
#> $estimate_args$impute_match_data[[2]]$Tr
#>  [1] 1 1 1 1 1 1 0 0 0 0 0 0
#> 
#> $estimate_args$impute_match_data[[2]]$X
#>         unit_1    unit_2    unit_3    unit_8    unit_4    unit_6
#> [1,] 0.3390333 0.3634129 0.2076023 0.2790739 0.4189477 0.3805735
#> [2,] 0.3179273 0.3951802 0.3069985 0.3069985 0.3951802 0.3951802
#> 
#> $estimate_args$impute_match_data[[2]]$orig.weighted.treated.nobs
#> [1] 6
#> 
#> 
#> $estimate_args$impute_match_data[[3]]
#> $estimate_args$impute_match_data[[3]]$Y
#>  [1] 102 105 120 108 100  94  96 130  80 130 130 130  80 130
#> 
#> $estimate_args$impute_match_data[[3]]$Tr
#>  [1] 1 1 1 1 1 1 1 0 0 0 0 0 0 0
#> 
#> $estimate_args$impute_match_data[[3]]$X
#>         unit_1    unit_2     unit_3     unit_8    unit_5    unit_7    unit_9
#> [1,] 0.1263758 0.4882738 0.05619221 0.08784167 0.1109228 0.4151219 0.1038343
#> [2,] 0.3208826 0.4514355 0.32088265 0.32088265 0.3208826 0.4514355 0.3208826
#> 
#> $estimate_args$impute_match_data[[3]]$orig.weighted.treated.nobs
#> [1] 7
#> 
#> 
#> 
#> $estimate_args$sigsqiw
#>        sigsqiw
#> unit_1    18.0
#> unit_2     4.5
#> unit_3    72.0
#> unit_8    18.0
#> unit_5     8.0
#> unit_7    18.0
#> unit_9     8.0
#> unit_4  1250.0
#> unit_6  1250.0
#> 
#> 
#> $model
#> Call:
#> MASS::polr(formula = as.factor(W) ~ X)
#> 
#> Coefficients:
#>         X 
#> 0.3698975 
#> 
#> Intercepts:
#>      1|2      2|3 
#> 2.173021 3.967827 
#> 
#> Residual Deviance: 16.97068 
#> AIC: 22.97068 
#> 
#> $propensity_scores
#>                1         2          3
#> unit_1 0.5345909 0.3390333 0.12637579
#> unit_2 0.1483133 0.3634129 0.48827384
#> unit_3 0.7362055 0.2076023 0.05619221
#> unit_8 0.6330845 0.2790739 0.08784167
#> unit_5 0.5711498 0.3179273 0.11092282
#> unit_7 0.1896979 0.3951802 0.41512190
#> unit_9 0.5891672 0.3069985 0.10383431
#> unit_4 0.2601697 0.4189477 0.32088265
#> unit_6 0.1679910 0.3805735 0.45143553
#> 
#> $impute_match_data
#> $impute_match_data[[1]]
#> $impute_match_data[[1]]$Y
#>  [1] 100  94  96 130  80 102 105 108 105 105
#> 
#> $impute_match_data[[1]]$Tr
#>  [1] 1 1 1 1 1 0 0 0 0 0
#> 
#> $impute_match_data[[1]]$X
#>         unit_5    unit_7    unit_9    unit_4    unit_6
#> [1,] 0.5711498 0.1896979 0.5891672 0.2601697 0.1679910
#> [2,] 0.5345909 0.1483133 0.6330845 0.1483133 0.1483133
#> 
#> $impute_match_data[[1]]$orig.weighted.treated.nobs
#> [1] 5
#> 
#> 
#> $impute_match_data[[2]]
#> $impute_match_data[[2]]$Y
#>  [1] 102 105 120 108 130  80 100  94  96  96  94  94
#> 
#> $impute_match_data[[2]]$Tr
#>  [1] 1 1 1 1 1 1 0 0 0 0 0 0
#> 
#> $impute_match_data[[2]]$X
#>         unit_1    unit_2    unit_3    unit_8    unit_4    unit_6
#> [1,] 0.3390333 0.3634129 0.2076023 0.2790739 0.4189477 0.3805735
#> [2,] 0.3179273 0.3951802 0.3069985 0.3069985 0.3951802 0.3951802
#> 
#> $impute_match_data[[2]]$orig.weighted.treated.nobs
#> [1] 6
#> 
#> 
#> $impute_match_data[[3]]
#> $impute_match_data[[3]]$Y
#>  [1] 102 105 120 108 100  94  96 130  80 130 130 130  80 130
#> 
#> $impute_match_data[[3]]$Tr
#>  [1] 1 1 1 1 1 1 1 0 0 0 0 0 0 0
#> 
#> $impute_match_data[[3]]$X
#>         unit_1    unit_2     unit_3     unit_8    unit_5    unit_7    unit_9
#> [1,] 0.1263758 0.4882738 0.05619221 0.08784167 0.1109228 0.4151219 0.1038343
#> [2,] 0.3208826 0.4514355 0.32088265 0.32088265 0.3208826 0.4514355 0.3208826
#> 
#> $impute_match_data[[3]]$orig.weighted.treated.nobs
#> [1] 7
#> 
#> 
#> 
#> $impute_mat_sorted
#>          1   2   3
#> unit_1 102 100 130
#> unit_2 105  94  80
#> unit_3 120  96 130
#> unit_8 108  96 130
#> unit_5 102 100 130
#> unit_7 105  94  80
#> unit_9 108  96 130
#> unit_4 105  94 130
#> unit_6 105  94  80
#> 

  multiMatch(Y,W,X,trimming=0,match_on="covariates")

#> no unit IDs supplied; unit_ids will be assigned generically

#> $results
#>         Param Trt1 Trt2   Estimate   Variance VarianceAI2016
#> 1 EY(2)-EY(1)    1    2 -10.666667   9.111111             NA
#> 2 EY(3)-EY(1)    1    3   6.666667 615.580247             NA
#> 3 EY(3)-EY(2)    2    3  17.333333 613.925926             NA
#> 
#> $analysis_idx
#> NULL
#> 
#> $mu
#>   Param Trt Estimate
#> 1 EY(1)   1 106.6667
#> 2 EY(2)   2  96.0000
#> 3 EY(3)   3 113.3333
#> 
#> $impute_mat
#>          1   2   3
#> unit_1 102 100 130
#> unit_2 105  94  80
#> unit_3 120  96 130
#> unit_4 105  94 130
#> unit_5 102 100 130
#> unit_6 105  94  80
#> unit_7 105  94  80
#> unit_8 108  96 130
#> unit_9 108  96 130
#> 
#> $estimate_args
#> $estimate_args$W
#> unit_1 unit_2 unit_3 unit_8 unit_5 unit_7 unit_9 unit_4 unit_6 
#>      1      1      1      1      2      2      2      3      3 
#> 
#> $estimate_args$X
#>        [,1]
#> unit_1  5.5
#> unit_2 10.6
#> unit_3  3.1
#> unit_8  4.4
#> unit_5  5.1
#> unit_7  9.8
#> unit_9  4.9
#> unit_4  8.7
#> unit_6 10.2
#> 
#> $estimate_args$Y
#> unit_1 unit_2 unit_3 unit_8 unit_5 unit_7 unit_9 unit_4 unit_6 
#>    102    105    120    108    100     94     96    130     80 
#> 
#> $estimate_args$N
#> [1] 9
#> 
#> $estimate_args$num_trts
#> [1] 3
#> 
#> $estimate_args$trt_levels
#> [1] 1 2 3
#> 
#> $estimate_args$N_per_trt
#> W
#> 1 2 3 
#> 4 3 2 
#> 
#> $estimate_args$num_contrasts
#> [1] 3
#> 
#> $estimate_args$orig_to_sorted
#> [1] 1 2 3 8 5 7 9 4 6
#> 
#> $estimate_args$sorted_to_orig
#> [1] 1 2 3 8 5 9 6 4 7
#> 
#> $estimate_args$unit_ids_unsorted
#> [1] "unit_1" "unit_2" "unit_3" "unit_4" "unit_5" "unit_6" "unit_7" "unit_8"
#> [9] "unit_9"
#> 
#> $estimate_args$unit_ids_sorted
#> [1] "unit_1" "unit_2" "unit_3" "unit_8" "unit_5" "unit_7" "unit_9" "unit_4"
#> [9] "unit_6"
#> 
#> $estimate_args$match_on
#> [1] "covariates"
#> 
#> $estimate_args$model_options
#> $estimate_args$model_options$reference_level
#> [1] 1
#> 
#> 
#> $estimate_args$M_matches
#> [1] 1
#> 
#> $estimate_args$J_var_matches
#> [1] 1
#> 
#> $estimate_args$Yiw
#>          1   2   3
#> unit_1 102 100 130
#> unit_2 105  94  80
#> unit_3 120  96 130
#> unit_8 108  96 130
#> unit_5 102 100 130
#> unit_7 105  94  80
#> unit_9 108  96 130
#> unit_4 105  94 130
#> unit_6 105  94  80
#> 
#> $estimate_args$mean_Yiw
#>        1        2        3 
#> 106.6667  96.0000 113.3333 
#> 
#> $estimate_args$Kiw
#>        Kiw
#> unit_1   1
#> unit_2   3
#> unit_3   0
#> unit_8   1
#> unit_5   1
#> unit_7   3
#> unit_9   2
#> unit_4   5
#> unit_6   2
#> 
#> $estimate_args$impute_match_data
#> $estimate_args$impute_match_data[[1]]
#> $estimate_args$impute_match_data[[1]]$Y
#>  [1] 100  94  96 130  80 102 105 108 105 105
#> 
#> $estimate_args$impute_match_data[[1]]$Tr
#>  [1] 1 1 1 1 1 0 0 0 0 0
#> 
#> $estimate_args$impute_match_data[[1]]$X
#>      unit_5 unit_7 unit_9 unit_4 unit_6
#> [1,]    5.1    9.8    4.9    8.7   10.2
#> [2,]    5.5   10.6    4.4   10.6   10.6
#> 
#> $estimate_args$impute_match_data[[1]]$orig.weighted.treated.nobs
#> [1] 5
#> 
#> 
#> $estimate_args$impute_match_data[[2]]
#> $estimate_args$impute_match_data[[2]]$Y
#>  [1] 102 105 120 108 130  80 100  94  96  96  94  94
#> 
#> $estimate_args$impute_match_data[[2]]$Tr
#>  [1] 1 1 1 1 1 1 0 0 0 0 0 0
#> 
#> $estimate_args$impute_match_data[[2]]$X
#>      unit_1 unit_2 unit_3 unit_8 unit_4 unit_6
#> [1,]    5.5   10.6    3.1    4.4    8.7   10.2
#> [2,]    5.1    9.8    4.9    4.9    9.8    9.8
#> 
#> $estimate_args$impute_match_data[[2]]$orig.weighted.treated.nobs
#> [1] 6
#> 
#> 
#> $estimate_args$impute_match_data[[3]]
#> $estimate_args$impute_match_data[[3]]$Y
#>  [1] 102 105 120 108 100  94  96 130  80 130 130 130  80 130
#> 
#> $estimate_args$impute_match_data[[3]]$Tr
#>  [1] 1 1 1 1 1 1 1 0 0 0 0 0 0 0
#> 
#> $estimate_args$impute_match_data[[3]]$X
#>      unit_1 unit_2 unit_3 unit_8 unit_5 unit_7 unit_9
#> [1,]    5.5   10.6    3.1    4.4    5.1    9.8    4.9
#> [2,]    8.7   10.2    8.7    8.7    8.7   10.2    8.7
#> 
#> $estimate_args$impute_match_data[[3]]$orig.weighted.treated.nobs
#> [1] 7
#> 
#> 
#> 
#> $estimate_args$sigsqiw
#>        sigsqiw
#> unit_1    18.0
#> unit_2     4.5
#> unit_3    72.0
#> unit_8    18.0
#> unit_5     8.0
#> unit_7    18.0
#> unit_9     8.0
#> unit_4  1250.0
#> unit_6  1250.0
#> 
#> 
#> $model
#> NULL
#> 
#> $propensity_scores
#> NULL
#> 
#> $impute_match_data
#> $impute_match_data[[1]]
#> $impute_match_data[[1]]$Y
#>  [1] 100  94  96 130  80 102 105 108 105 105
#> 
#> $impute_match_data[[1]]$Tr
#>  [1] 1 1 1 1 1 0 0 0 0 0
#> 
#> $impute_match_data[[1]]$X
#>      unit_5 unit_7 unit_9 unit_4 unit_6
#> [1,]    5.1    9.8    4.9    8.7   10.2
#> [2,]    5.5   10.6    4.4   10.6   10.6
#> 
#> $impute_match_data[[1]]$orig.weighted.treated.nobs
#> [1] 5
#> 
#> 
#> $impute_match_data[[2]]
#> $impute_match_data[[2]]$Y
#>  [1] 102 105 120 108 130  80 100  94  96  96  94  94
#> 
#> $impute_match_data[[2]]$Tr
#>  [1] 1 1 1 1 1 1 0 0 0 0 0 0
#> 
#> $impute_match_data[[2]]$X
#>      unit_1 unit_2 unit_3 unit_8 unit_4 unit_6
#> [1,]    5.5   10.6    3.1    4.4    8.7   10.2
#> [2,]    5.1    9.8    4.9    4.9    9.8    9.8
#> 
#> $impute_match_data[[2]]$orig.weighted.treated.nobs
#> [1] 6
#> 
#> 
#> $impute_match_data[[3]]
#> $impute_match_data[[3]]$Y
#>  [1] 102 105 120 108 100  94  96 130  80 130 130 130  80 130
#> 
#> $impute_match_data[[3]]$Tr
#>  [1] 1 1 1 1 1 1 1 0 0 0 0 0 0 0
#> 
#> $impute_match_data[[3]]$X
#>      unit_1 unit_2 unit_3 unit_8 unit_5 unit_7 unit_9
#> [1,]    5.5   10.6    3.1    4.4    5.1    9.8    4.9
#> [2,]    8.7   10.2    8.7    8.7    8.7   10.2    8.7
#> 
#> $impute_match_data[[3]]$orig.weighted.treated.nobs
#> [1] 7
#> 
#> 
#> 
#> $impute_mat_sorted
#>          1   2   3
#> unit_1 102 100 130
#> unit_2 105  94  80
#> unit_3 120  96 130
#> unit_8 108  96 130
#> unit_5 102 100 130
#> unit_7 105  94  80
#> unit_9 108  96 130
#> unit_4 105  94 130
#> unit_6 105  94  80
#>