Matching on GPS with multilevel treatments

Matching on GPS with multilevel treatments

multilevelGPSMatch(Y, W, X, Trimming, GPSM = "multinomiallogisticReg")

Arguments

Y
A continuous response vector (1 x n)

W
A treatment vector (1 x n) with numerical values indicating treatment groups

X
A covariate matrix (p x n) with no intercept. When GPSM="existing", then X must be a vector (1 x n) of user-specified propensity scores.

Trimming
An indicator of whether trimming the sample to ensure overlap

GPSM
An indicator of the methods used for estimating GPS, options include "multinomiallogisticReg", "ordinallogisticReg" for proportional odds or cumulative logit model, and "existing" for user-specified propensity score via the parameter X. Defaults to "multinomiallogisticReg"

Value

A list element including:

tauestimate: A vector of estimates for pairwise treatment effects

varestimate: A vector of variance estimates for tauestimate, using Abadie & Imbens (2006)'s method

varestimateAI2012: A vector of variance estimates for tauestimate, when matching on the generalized propensity score, using Abadie & Imbens (2016)'s method. This variance estimate takes into account of the uncertainty in estimating the GPS. This variable is named AI2012 (not AI2016) for backwards compatibility.

analysis_idx: a list containing the indices_kept (analyzed) and indices_dropped (trimmed) based on Crump et al. (2009)'s method.

References

Yang, S., Imbens G. W., Cui, Z., Faries, D. E., & Kadziola, Z. (2016) Propensity Score Matching and Subclassification in Observational Studies with Multi-Level Treatments. Biometrics, 72, 1055-1065. https://doi.org/10.1111/biom.12505

Abadie, A., & Imbens, G. W. (2006). Large sample properties of matching estimators for average treatment effects. Econometrica, 74(1), 235-267. https://doi.org/10.1111/j.1468-0262.2006.00655.x

Abadie, A., & Imbens, G. W. (2016). Matching on the estimated propensity score. Econometrica, 84(2), 781-807. https://doi.org/10.3982/ECTA11293

Crump, R. K., Hotz, V. J., Imbens, G. W., & Mitnik, O. A. (2009). Dealing with limited overlap in estimation of average treatment effects. Biometrika, 96(1), 187-199. https://doi.org/10.1093/biomet/asn055

See also

Examples

  X <- c(5.5,10.6,3.1,8.7,5.1,10.2,9.8,4.4,4.9)
  Y <- c(102,105,120,130,100,80,94,108,96)
  W <- c(1,1,1,3,2,3,2,1,2)
  multilevelGPSMatch(Y,W,X,Trimming=0,GPSM="multinomiallogisticReg")

#> $tauestimate
#> EY(2)-EY(1) EY(3)-EY(1) EY(3)-EY(2) 
#>  -10.444444    6.666667   17.111111 
#> 
#> $varestimate
#> EY(2)-EY(1) EY(3)-EY(1) EY(3)-EY(2) 
#>    8.545953  616.913580  611.122085 
#> 
#> $varestimateAI2012
#> EY(2)-EY(1) EY(3)-EY(1) EY(3)-EY(2) 
#>    8.302024  411.456234  434.247037 
#> 
#> $analysisidx
#> [1] 1 2 3 4 5 6 7 8 9
#> 

  multilevelGPSMatch(Y,W,X,Trimming=1,GPSM="multinomiallogisticReg")

#> $tauestimate
#> EY(2)-EY(1) EY(3)-EY(1) EY(3)-EY(2) 
#>      -9.375       5.875      15.250 
#> 
#> $varestimate
#> EY(2)-EY(1) EY(3)-EY(1) EY(3)-EY(2) 
#>    7.794922  582.654297  576.304688 
#> 
#> $varestimateAI2012
#> EY(2)-EY(1) EY(3)-EY(1) EY(3)-EY(2) 
#>    5.072057  383.848575  430.978089 
#> 
#> $analysisidx
#> 1 2 4 5 6 7 8 9 
#> 1 2 4 5 6 7 8 9 
#>