Matching on X with multilevel treatments

Matching on X with multilevel treatments

multilevelMatchX(Y, W, X)

Arguments

Y
A continuous response vector (1 x n)

W
A treatment vector (1 x n) with numerical values indicating treatment groups

X
A covariate matrix (p x n) with no intercept

Value

A list with 2 elements: tauestimate, varestimate, where tauestimate is a vector of estimates for pairwise treatment effects, and varestimate is a vector of variance estimates for tauestimate, using Abadie & Imbens (2006)'s method.

References

Yang, S., Imbens G. W., Cui, Z., Faries, D. E., & Kadziola, Z. (2016) Propensity Score Matching and Subclassification in Observational Studies with Multi-Level Treatments. Biometrics, 72, 1055-1065. https://doi.org/10.1111/biom.12505

Abadie, A., & Imbens, G. W. (2006). Large sample properties of matching estimators for average treatment effects. econometrica, 74(1), 235-267. https://doi.org/10.1111/j.1468-0262.2006.00655.x

Abadie, A., & Imbens, G. W. (2016). Matching on the estimated propensity score. Econometrica, 84(2), 781-807. https://doi.org/10.3982/ECTA11293

Crump, R. K., Hotz, V. J., Imbens, G. W., & Mitnik, O. A. (2009). Dealing with limited overlap in estimation of average treatment effects. Biometrika, 96(1), 187-199. https://doi.org/10.1093/biomet/asn055

See also

Examples

  X<-c(5.5,10.6,3.1,8.7,5.1,10.2,9.8,4.4,4.9)
  Y<-c(102,105,120,130,100,80,94,108,96)
  W<-c(1,1,1,3,2,3,2,1,2)
  multilevelMatchX(Y,W,X)

#> $tauestimate
#> EY(2)-EY(1) EY(3)-EY(1) EY(3)-EY(2) 
#>  -10.666667    6.666667   17.333333 
#> 
#> $varestimate
#> EY(2)-EY(1) EY(3)-EY(1) EY(3)-EY(2) 
#>    9.111111  615.580247  613.925926 
#>