Calculate imputed potential outcomes and other estimation components

This is a plumbing function, called from matchImputePO; they output the same list of information.

wrangleImputations(match_output, M_matches, Y, which_same_trt,
  which_diff_trt)

Arguments

match_output
The output from Match when matching different treatment levels to impute potential outcomes (for point estimation of causal estimands).

M_matches
Number of matches per unit for imputing potential outcomes, as in Abadie and Imbens (2006).

Y
A response vector (1 x n)

which_same_trt
Numeric vector providing indices of individuals observed to have treatment level kk

which_diff_trt
Numeric vector providing indices of individuals observed to have treatment level NOT equal to kk

Value

A list, many of which are necessary arguments for estimateTau:

Yiw The fully-imputed data set of all potential outcomes.

mean_Yiw Mean of Yiw across all units i. These are the estimates of the population-level mean estimands. See estimateTau.

Kiw The vector of number of times unit i used as a match in imputation.

impute_match_data Extra information from the main matching procedure.