Stratification on GPS with multilevel treatments

Stratification on GPS with multilevel treatments

multilevelGPSStratification(Y, W, X, NS, GPSM = "multinomiallogisticReg",
  linearp = 0, nboot)

Arguments

Y
A continuous response vector (1 x n)

W
A treatment vector (1 x n) with numerical values indicating treatment groups

X
A covariate matrix (p x n) with no intercept

NS
The number of strata: (only required in the function multilevelGPSStratification)

GPSM
An indicator of the methods used for estimating GPS, options include "multinomiallogisticReg", "ordinallogisticReg", and "existing"

linearp
An indicator of subclassification on GPS (=0) or linear predictor of GPS (=1): (only required in the function multilevelGPSStratification)

nboot
The number of boot replicates for variance estimation: (only required in the function multilevelGPSStratification)

Value

A list with two elements, tauestimate, varestimate, where tauestimate is a vector of estimates for pairwise treatment effects, and varestimate is a vector of variance estimates, using bootstrapping method.

References

Yang, S., Imbens G. W., Cui, Z., Faries, D. E., & Kadziola, Z. (2016) Propensity Score Matching and Subclassification in Observational Studies with Multi-Level Treatments. Biometrics, 72, 1055-1065. https://doi.org/10.1111/biom.12505

Abadie, A., & Imbens, G. W. (2006). Large sample properties of matching estimators for average treatment effects. Econometrica, 74(1), 235-267. https://doi.org/10.1111/j.1468-0262.2006.00655.x

Abadie, A., & Imbens, G. W. (2016). Matching on the estimated propensity score. Econometrica, 84(2), 781-807. https://doi.org/10.3982/ECTA11293

See also

Examples

simulated_data <- multilevelMatching::simulated_data
set.seed(123)
multilevelGPSStratification(
  Y = simulated_data$outcome ,
  W = simulated_data$treatment,
  X = simulated_data[ ,names(simulated_data) %in% paste0("covar", 1:6)],
  GPSM = "multinomiallogisticReg",
  NS = 5,
  linearp = TRUE,
  nboot = 10
)

#> $tauestimate
#> EY(2)-EY(1) EY(3)-EY(1) EY(3)-EY(2) 
#>  -0.3038424   0.4201045   0.7239469 
#> 
#> $varestimate
#> EY(2)-EY(1) EY(3)-EY(1) EY(3)-EY(2) 
#>   0.1658798   0.2881434   0.6232490 
#>